refactor: rename mainPage directory to 'main-page'
