+++ src/app/main-page/PromotionCard.tsx
