+++ src/app/main-page/styles.ts
